chore: Corrijo el tipo de total en Sale y agrego los dtos de create

+++ src/main/java/com/gyl/CrudGyl/sale/entity/Sale.java
@@ -21,5 +21,5 @@ public class Sale {
     Instant date;
 
     @Column(name = "total")
-    Long total;
+    Double total;
 }
